flask-server/requirements.txt
attrs==21.2.0
cffi==1.15.0
click==8.0.1
cryptography==36.0.0
ddtrace==1.5.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.0
pyparsing==2.4.7
six==1.16.0
tenacity==8.0.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
